pdfbot/requirements.txt
python-telegram-bot==22.6
pdf2image
pillow
img2pdf
PyPDF2
python-docx
